renames util folders to utils folders
