+++ server/index.js
@@ -2,9 +2,9 @@ const webpack = require('webpack')
 const express = require('express')
 const bodyParser = require('body-parser')
 const cron = require('node-cron')
-const { PORT, inProduction } = require('@util/common')
-const routes = require('@util/routes')
-const logger = require('@util/logger')
+const routes = require('@utils/routes')
+const logger = require('@utils/logger')
+const { PORT, inProduction } = require('@utils/common')
 
 const processNewCompletions = require('./scripts/processNewCompletions')
 const processOldCompletions = require('./scripts/processOldCompletions')
