+++ server/scripts/processCSV.js
@@ -4,7 +4,7 @@ const {
   isValidOodiDate,
   isValidGrade,
   isValidCreditAmount,
-} = require('../../util/validators')
+} = require('../../utils/validators')
 
 const LANGUAGES = {
   fi: 1,
